Start of vocab update
There are lots of issues that are blocked by the vocab page not being up
to scratch. A first attempt and some thinking:

Should the version changes document (#77) be part of the vocab? It is
about mapping changes to vocab and we haven't done a good job of keeping
the  common structures document up to date.

This is a first attempt to start to keep extensions and values apart
from each other #48 but I am still thinking about ways of making it
explicit.

data types, value spaces and sample values are an attempt to help define
'what it is which sits' : #51.

I'm not actually clear about the the common statements page #52 either.
Is this something that should be merged in to vocab to help structure
that document and remind us to keep on top of it?

Any feelings on any of these approaches to documentation @arc12 ?